Do not greedily consume all the client's stdin. Instead, read from the client's stdin only when something reads from `System.in` in the build process.
This avoids buffering potentially large amounts of unconsumed stdin, and allows the client to use stdin for other purposes, for example prompting for a secret or using an interactive TUI for user prompting.

It will also allow the rich console to render the status bar without leaving garbage behind when stdin is used.